fix(pagination): do not apply offset on the root query when a subquery is present
