+++ tests/integration/conftest.py
@@ -1,13 +1,35 @@
 from __future__ import annotations
 
-from .fixtures import async_engine, async_session, asyncpg_engine, engine, psycopg_async_engine, psycopg_engine, session
+from .fixtures import (
+    any_query,
+    async_engine,
+    async_session,
+    asyncpg_engine,
+    engine,
+    no_session_query,
+    psycopg_async_engine,
+    psycopg_engine,
+    raw_colors,
+    raw_fruits,
+    raw_users,
+    seed_db_async,
+    seed_db_sync,
+    session,
+)
 
 __all__ = (
+    "any_query",
     "async_engine",
     "async_session",
     "asyncpg_engine",
     "engine",
+    "no_session_query",
     "psycopg_async_engine",
     "psycopg_engine",
+    "raw_colors",
+    "raw_fruits",
+    "raw_users",
+    "seed_db_async",
+    "seed_db_sync",
     "session",
 )
